Thrall already setup to run two streams and merge using akka. removed unused imports and reorganized kinesis folder so all kinesis files are together. added a question comment for sam in thrall config but will remove.
